:hammer: Improve array formatting and error handling in format.sh script
This commit refactors the script to append square brackets to the beginning
and end of the array and passes it through the jq processor. If jq returns
a success status code, the temporary output file will be moved to the
specified file. Otherwise, an error message will be printed and the
temporary file will be deleted.